# oapispec/swagger.yaml
basePath: /
definitions:
...
  admin.AdminCreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    - permissions
    type: object
  admin.AdminCreateRoleResponse:
    properties:
      message:
        type: string
      role:
        $ref: '#/definitions/db.Role'
      success:
        type: boolean
    type: object
...
  admin.AdminDeleteRoleResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  admin.AdminUpdateOrgRequest:
    properties:
      description:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
    required:
    - name
    - settings
    type: object
  admin.AdminUpdateOrgResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  admin.AdminUpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    - permissions
    type: object
  admin.AdminUpdateRoleResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  db.GetMinimalRolesForOrgRow:
    properties:
      createdAt:
        type: string
      id:
        type: string
      roleDesc:
        type: string
      roleName:
        type: string
      updatedAt:
        type: string
    type: object
  db.Org:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  db.OrgDomain:
    properties:
      autoJoin:
        type: boolean
      autoJoinRoleId:
        type: string
      autoJoinRoleName:
        type: string
      createdAt:
        type: string
      domain:
        type: string
      orgId:
        type: string
      updatedAt:
        type: string
      verified:
        type: boolean
      verifiedAt:
        type: string
    type: object
  db.Role:
    properties:
      createdAt:
        type: string
      id:
        type: string
      orgId:
        type: string
      permissions:
        items:
          type: string
        type: array
      roleDesc:
        type: string
      roleName:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.OrgMFASettings:
    properties:
      required:
        type: boolean
      roles:
        additionalProperties:
          type: string
        description: role name to role-id mapping
        type: object
    type: object
  models.OrgSettings:
    properties:
      mfa:
        $ref: '#/definitions/models.OrgMFASettings'
    type: object
  reserr.ErrorResponse:
    properties:
      debug:
        description: DebugMessage is a technical message that can be used for debugging.
        type: string
      message:
        description: Message is a user-friendly message that can be displayed to the
          end user.
        type: string
    type: object
  resquery.FilterMode:
    enum:
    - AND
    - OR
    type: string
    x-enum-varnames:
    - FilterModeAND
    - FilterModeOR
  resquery.FilterOption:
    properties:
      field:
        type: string
      op:
        $ref: '#/definitions/resquery.QueryFilterOp'
      value: {}
    type: object
  resquery.QueryFilterOp:
    enum:
    - contains
    - equals
    - lte
    - gte
    - lt
    - gt
    type: string
    x-enum-varnames:
    - OpContains
    - OpEquals
    - OpLte
    - OpGte
    - OpLt
    - OpGt
  resquery.QueryFilters:
    properties:
      mode:
        $ref: '#/definitions/resquery.FilterMode'
      options:
        items:
          $ref: '#/definitions/resquery.FilterOption'
        minItems: 1
        type: array
    required:
    - mode
    - options
    type: object
  resquery.QueryPagination:
    properties:
      page:
        description: Page is the page number to retrieve (0-based).
        minimum: 0
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - page
    - pageSize
    type: object
  resquery.SortOption:
    properties:
      direction:
        enum:
        - asc
        - desc
        - ASC
        - DESC
        type: string
      field:
        type: string
    required:
    - direction
    - field
    type: object
  sys_admin.SysAdminCreateOrgRequest:
    properties:
      avatarURL:
        type: string
      description:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/models.OrgSettings'
      slug:
        type: string
    required:
    - name
    - settings
    - slug
    type: object
  sys_admin.SysAdminCreateOrgResponse:
    properties:
      message:
        type: string
      orgId:
        type: string
      success:
        type: boolean
    type: object
  sys_admin.SysAdminDeleteOrgResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  sys_admin.SysAdminGetOrgDetailsResponse:
    properties:
      domains:
        items:
          $ref: '#/definitions/db.OrgDomain'
        type: array
      message:
        type: string
      org:
        $ref: '#/definitions/db.Org'
      roles:
        items:
          $ref: '#/definitions/db.GetMinimalRolesForOrgRow'
        type: array
      success:
        type: boolean
    type: object
  sys_admin.SysAdminListOrgsRequest:
    properties:
      filters:
        $ref: '#/definitions/resquery.QueryFilters'
      pagination:
        $ref: '#/definitions/resquery.QueryPagination'
      sort:
        items:
          $ref: '#/definitions/resquery.SortOption'
        type: array
    required:
    - sort
    type: object
  sys_admin.SysAdminListOrgsResponse:
    properties:
      message:
        type: string
      orgs:
        description: List of organizations
        items:
          $ref: '#/definitions/db.Org'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/resquery.QueryPagination'
        description: Pagination details, default if not provided in the request
      success:
        type: boolean
      total:
        description: Total number of organizations matching the filters (if any)
        type: integer
    type: object
...
host: localhost:3360
info:
  contact:
    email: contact@nbrglm.com
    name: NBRGLM Support
    url: https://nbrglm.com/support
  description: This is the NBRGLM Nexeres API Spec documentation.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://nbrglm.com/nexeres/terms
  title: NBRGLM Nexeres API Spec
  version: 0.0.1
paths:
...
  /api/admin/orgs/{orgId}:
    post:
      consumes:
      - application/json
      description: Handles AdminUpdateOrg requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUpdateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminUpdateOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminUpdateOrg Endpoint
      tags:
      - admin
...
  /api/admin/orgs/{orgId}/roles:
    put:
      consumes:
      - application/json
      description: Handles CreateRole requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminCreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminCreateRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: CreateRole Endpoint
      tags:
      - admin
  /api/admin/orgs/{orgId}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Handles AdminDeleteRole requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminDeleteRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminDeleteRole Endpoint
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Handles AdminUpdateRole requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminUpdateRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminUpdateRole Endpoint
      tags:
      - admin
...
  /api/sys/admin/orgs:
    post:
      consumes:
      - application/json
      description: Handles SysAdminListOrgs requests
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys_admin.SysAdminListOrgsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminListOrgsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminListOrgs Endpoint
      tags:
      - sys_admin
    put:
      consumes:
      - application/json
      description: Handles SysAdminCreateOrg requests
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sys_admin.SysAdminCreateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminCreateOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminCreateOrg Endpoint
      tags:
      - sys_admin
  /api/sys/admin/orgs/{orgId}:
    delete:
      consumes:
      - application/json
      description: Handles SysAdminDeleteOrg requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminDeleteOrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminDeleteOrg Endpoint
      tags:
      - sys_admin
    get:
      consumes:
      - application/json
      description: Handles SysAdminGetOrgDetails requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sys_admin.SysAdminGetOrgDetailsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: SysAdminGetOrgDetails Endpoint
      tags:
      - sys_admin
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-NEXERES-API-Key
    type: apiKey
  AdminHeaderAuth:
    in: header
    name: X-NEXERES-Admin-Token
    type: apiKey
  RefreshHeaderAuth:
    in: header
    name: X-NEXERES-Refresh-Token
    type: apiKey
  SessionHeaderAuth:
    in: header
    name: X-NEXERES-Session-Token
    type: apiKey
swagger: "2.0"